lambda/requirements.txt
requests
amazon-textract-caller
amazon-textract-response-parser
amazon-textract-prettyprinter
